Fix a command injection vulnerability in vanilla Minecraft 20w21a -> 23w04a
Fix a command injection vulnerability allowing malicious resource pack to unexpectedly execute code on Linux clients running vanilla Minecraft 1.16 (20w21a) -> 1.19.3 (23w04a).

It was recently found that the Tiny File Dialogs library is vulnerable to command injection techniques on Linux when invoked with untrusted data. This fix makes a small change in the create world `MoreOptionsDialog` screen to pass a hardcoded string as the dialog title.

Many thanks to ThatGravyBoat and Moulberry for investigating this issue and reporting it to us in confidence.